Move cart folder inside sections
